change UpdateTask func

+++ models.todolist.go
@@ -1,7 +1,7 @@
 package main
 
 type task struct {
-	TaskID  uint16 `gorm:"primary_key:yes;column:taskID"`
+	TaskID  int    `gorm:"primary_key:yes;column:taskID"`
 	Title   string `form:"title" json:"title" xml:"title"  binding:"required"`
 	Content string `form:"content" json:"content" xml:"content" binding:"required"`
 	Done    bool   `form:"done" json:"done" xml:"done"`
